Adds Dockerfile, corrects build, other improvements
Closes #17
Updates Markdown editor library to a new version, because the previous one was breaking the production build. Adds the build step to the build workflow.
Adds a web request when the admin UI is loaded, to validate the access token in the sessionStorage. This reduces performance, but prevents displaying the admin UI to users who add any value to their session storage (note: API calls validate access tokens, but static assets would be displayed). It is probably debatable whether this web request to validate the access token in the sessionStorage should take place or not.